requirements.txt
langchain==0.3.14
langchain-openai==1.1.12
langchain-community==0.3.14
langchain-chroma==1.1.0
langchain-text-splitters==0.3.4
chromadb==0.5.23
streamlit==1.41.1
python-dotenv==1.0.1
